Moved twitter examples into apps, and the twitter helpers to lib, added lib to sys paths.
